refactor(forms): restructure files
