+++ packages/core/forms/src/generator/fields/FieldAdvanced.vue
@@ -26,7 +26,7 @@
 </template>
 
 <script>
-import abstractField from '../abstractField'
+import abstractField from './abstractField'
 
 export default {
   mixins: [abstractField],
