+++ packages/core/forms/src/generator/fields/FieldAutoSuggest.vue
@@ -41,10 +41,10 @@
 </template>
 
 <script>
-import { FORMS_API_KEY } from '../../../const'
-import abstractField from '../abstractField'
+import { FORMS_API_KEY } from '../../const'
+import abstractField from './abstractField'
 import debounce from 'lodash-es/debounce'
-import { isValidUuid } from '../../utils/isValidUuid'
+import { isValidUuid } from '../utils/isValidUuid'
 
 const requestResultsLimit = 50
 const fieldStates = {
