+++ packages/core/forms/src/generator/fields/FieldMultiselect.vue
@@ -13,7 +13,7 @@
 </template>
 
 <script>
-import abstractField from '../abstractField'
+import abstractField from './abstractField'
 
 export default {
   mixins: [abstractField],
