+++ packages/core/forms/src/generator/fields/FieldObject.vue
@@ -71,7 +71,7 @@
 </template>
 
 <script>
-import abstractField from '../abstractField'
+import abstractField from './abstractField'
 
 export default {
   mixins: [abstractField],
