+++ packages/core/forms/src/generator/fields/FieldRadio.vue
@@ -27,7 +27,7 @@
 </template>
 
 <script>
-import abstractField from '../abstractField'
+import abstractField from './abstractField'
 
 export default {
   mixins: [abstractField],
