+++ packages/core/forms/src/generator/fields/FieldSelect.vue
@@ -13,7 +13,7 @@ select.form-control(v-model="value", :disabled="disabled || null", :name="schema
 import isObject from 'lodash-es/isObject'
 import isNil from 'lodash-es/isNil'
 import find from 'lodash-es/find'
-import abstractField from '../abstractField'
+import abstractField from './abstractField'
 
 export default {
   mixins: [abstractField],
