+++ packages/core/forms/src/generator/fields/FieldSelectionGroup.vue
@@ -49,7 +49,7 @@
 </template>
 
 <script>
-import abstractField from '../abstractField'
+import abstractField from './abstractField'
 
 export default {
   mixins: [abstractField],
