+++ packages/core/forms/src/generator/fields/FieldSwitch.vue
@@ -6,7 +6,7 @@ label
 </template>
 
 <script>
-import abstractField from '../abstractField'
+import abstractField from './abstractField'
 
 export default {
   mixins: [abstractField],
