+++ packages/core/forms/src/generator/fields/FieldTextArea.vue
@@ -15,7 +15,7 @@ textarea.form-control(
 </template>
 
 <script>
-import abstractField from '../abstractField'
+import abstractField from './abstractField'
 
 export default {
   mixins: [abstractField],
