+++ packages/core/forms/src/generator/utils/fieldsLoader.ts
@@ -1,35 +1,20 @@
-export { default as FieldCheckbox } from '../fields/core/fieldCheckbox.vue'
-export { default as FieldChecklist } from '../fields/core/fieldChecklist.vue'
-export { default as FieldInput } from '../fields/core/FieldInput.vue'
-export { default as FieldLabel } from '../fields/core/FieldLabel.vue'
-export { default as FieldRadios } from '../fields/core/fieldRadios.vue'
-export { default as FieldSelect } from '../fields/core/fieldSelect.vue'
-export { default as FieldSubmit } from '../fields/core/fieldSubmit.vue'
-export { default as FieldTextArea } from '../fields/core/fieldTextArea.vue'
-export { default as FieldUpload } from '../fields/core/fieldUpload.vue'
-export { default as FieldCleave } from '../fields/optional/fieldCleave.vue'
-export { default as FieldDateTimePicker } from '../fields/optional/fieldDateTimePicker.vue'
-export { default as FieldGoogleAddress } from '../fields/optional/fieldGoogleAddress.vue'
-export { default as FieldImage } from '../fields/optional/fieldImage.vue'
-export { default as FieldMasked } from '../fields/optional/fieldMasked.vue'
-export { default as FieldNoUiSlider } from '../fields/optional/fieldNoUiSlider.vue'
-export { default as FieldPikaday } from '../fields/optional/fieldPikaday.vue'
-export { default as FieldRangeSlider } from '../fields/optional/fieldRangeSlider.vue'
-export { default as FieldSelectEx } from '../fields/optional/fieldSelectEx.vue'
-export { default as FieldSpectrum } from '../fields/optional/fieldSpectrum.vue'
-export { default as FieldStaticMap } from '../fields/optional/fieldStaticMap.vue'
-export { default as fieldSwitch } from '../fields/optional/fieldSwitch.vue'
-
-export { default as FieldObject } from '../fields/advanced/FieldObject.vue'
-export { default as FieldObjectAdvanced } from '../fields/advanced/FieldObjectAdvanced.vue'
-export { default as FieldArray } from '../fields/advanced/FieldArray.vue'
-export { default as FieldMetric } from '../fields/advanced/FieldMetric.vue'
-export { default as FieldArrayItem } from '../fields/advanced/FieldArrayItem.vue'
-export { default as FieldArrayMultiItem } from '../fields/advanced/FieldArrayMultiItem.vue'
-export { default as FieldAdvanced } from '../fields/advanced/FieldAdvanced.vue'
-export { default as FieldAutoSuggest } from '../fields/advanced/FieldAutoSuggest.vue'
-export { default as FieldSelectionGroup } from '../fields/advanced/FieldSelectionGroup.vue'
-export { default as FieldRadio } from '../fields/advanced/FieldRadio.vue'
-export { default as FieldArrayCardContainer } from '../fields/advanced/FieldArrayCardContainer.vue'
-export { default as FieldMultiselect } from '../fields/advanced/FieldMultiselect.vue'
-export { default as FieldKeyValuePairs } from '../fields/advanced/FieldKeyValuePairs.vue'
+export { default as FieldAdvanced } from '../fields/FieldAdvanced.vue'
+export { default as FieldArray } from '../fields/FieldArray.vue'
+export { default as FieldArrayCardContainer } from '../fields/FieldArrayCardContainer.vue'
+export { default as FieldArrayItem } from '../fields/FieldArrayItem.vue'
+export { default as FieldArrayMultiItem } from '../fields/FieldArrayMultiItem.vue'
+export { default as FieldAutoSuggest } from '../fields/FieldAutoSuggest.vue'
+export { default as FieldCheckbox } from '../fields/FieldCheckbox.vue'
+export { default as FieldChecklist } from '../fields/FieldChecklist.vue'
+export { default as FieldInput } from '../fields/FieldInput.vue'
+export { default as FieldKeyValuePairs } from '../fields/FieldKeyValuePairs.vue'
+export { default as FieldLabel } from '../fields/FieldLabel.vue'
+export { default as FieldMetric } from '../fields/FieldMetric.vue'
+export { default as FieldMultiselect } from '../fields/FieldMultiselect.vue'
+export { default as FieldObject } from '../fields/FieldObject.vue'
+export { default as FieldObjectAdvanced } from '../fields/FieldObjectAdvanced.vue'
+export { default as FieldRadio } from '../fields/FieldRadio.vue'
+export { default as FieldSelect } from '../fields/FieldSelect.vue'
+export { default as FieldSelectionGroup } from '../fields/FieldSelectionGroup.vue'
+export { default as FieldSwitch } from '../fields/FieldSwitch.vue'
+export { default as FieldTextArea } from '../fields/FieldTextArea.vue'
